stock-prediction-engine/requirements.txt
# Stock Prediction Engine - Phase 3: News Intelligence Engine
# Requirements for comprehensive news scraping and sentiment analysis system

# Web Scraping & HTTP
requests>=2.31.0
beautifulsoup4>=4.12.2
scrapy>=2.11.0
selenium>=4.15.0
webdriver-manager>=4.0.1
fake-useragent>=1.4.0
lxml>=4.9.3
html5lib>=1.1
urllib3>=2.0.7

# News APIs & Social Media
newsapi-python>=0.2.7
alpha-vantage>=2.3.1
praw>=7.7.1
tweepy>=4.14.0
feedparser>=6.0.10

# AI & Natural Language Processing
openai>=1.3.0
spacy>=3.7.0
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2

# Data Processing & Analysis
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
alembic>=1.12.0
redis>=4.6.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
configparser>=5.3.0
python-decouple>=3.8

# Logging & Monitoring
loguru>=0.7.2
structlog>=23.2.0
prometheus-client>=0.17.0
sentry-sdk>=1.38.0

# Date & Time Handling
python-dateutil>=2.8.2
pytz>=2023.3
arrow>=1.3.0

# Async & Concurrency
asyncio
aiohttp>=3.8.0
asyncpg>=0.28.0
celery>=5.3.0
redis>=4.6.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0
factory-boy>=3.3.0
responses>=0.23.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0
pydantic>=2.4.0

# Data Validation & Serialization
marshmallow>=3.20.0
jsonschema>=4.19.0
cerberus>=1.3.5

# Caching
cachetools>=5.3.0
diskcache>=5.6.3

# File Processing
openpyxl>=3.1.0
python-docx>=0.8.11
PyPDF2>=3.0.0
chardet>=5.2.0

# Networking & Security
cryptography>=41.0.0
certifi>=2023.7.22
urllib3>=2.0.7
requests[security]>=2.31.0

# Job Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Progress Bars & CLI
tqdm>=4.66.0
click>=8.1.0
rich>=13.6.0
typer>=0.9.0

# Financial Data (for correlation analysis)
yfinance>=0.2.0
pandas-datareader>=0.10.0
alpha-vantage>=2.3.1

# Text Processing & Cleaning
regex>=2023.10.0
ftfy>=6.1.0
unidecode>=1.3.0
langdetect>=1.0.9

# Machine Learning Utilities
joblib>=1.3.0
pickle5>=0.0.12
dill>=0.3.7

# Development Tools
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.6.0
pre-commit>=3.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Memory & Performance
memory-profiler>=0.61.0
psutil>=5.9.0
py-spy>=0.3.14

# Browser Automation (for advanced scraping)
playwright>=1.40.0
pyppeteer>=1.0.2

# Email & Notifications
emails>=0.6.0
twilio>=8.10.0
slack-sdk>=3.23.0

# Geographic & Location Data
geopy>=2.4.0
pycountry>=22.3.13

# Image Processing (for news images)
Pillow>=10.0.0
opencv-python>=4.8.0

# XML Processing
xmltodict>=0.13.0
defusedxml>=0.7.1

# Timezone Handling
babel>=2.13.0
pytz>=2023.3

# HTTP Client Enhancements
httpx>=0.25.0
aiofiles>=23.2.0

# Data Export Formats
xlsxwriter>=3.1.0
tabulate>=0.9.0

# Monitoring & Health Checks
healthcheck>=1.3.3
watchdog>=3.0.0

# Rate Limiting
ratelimit>=2.2.1
slowapi>=0.1.9

# Security & Authentication
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.0

# Environment Detection
python-dotenv>=1.0.0
environs>=10.0.0

# Serialization
orjson>=3.9.0
msgpack>=1.0.7

# Command Line Interface
fire>=0.5.0
python-fire>=0.5.0

# Time Series Analysis
statsmodels>=0.14.0
ta-lib>=0.4.0

# Data Streaming
kafka-python>=2.0.2
pika>=1.3.0

# Utility Libraries
more-itertools>=10.1.0
toolz>=0.12.0
funcy>=2.0.0

# Financial Libraries
quantlib>=1.32.0
financedatabase>=2.2.0

# Spell Checking & Text Correction
pyspellchecker>=0.7.0
autocorrect>=2.6.0

# Language Detection
polyglot>=16.7.4
langdetect>=1.0.9

# Fuzzy Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# HTML/XML Parsing
selectolax>=0.3.0
pyquery>=2.0.0

# User Agent Rotation
fake-useragent>=1.4.0
user-agent>=0.1.10

# Proxy Support
requests[socks]>=2.31.0
PySocks>=1.7.1

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Workflow Management
prefect>=2.14.0
dagster>=1.5.0

# Code Quality
bandit>=1.7.5
safety>=2.3.0

# Development Dependencies (optional, for development environment)
jupyter>=1.0.0
ipython>=8.16.0
ipykernel>=6.25.0
notebook>=7.0.0

# Documentation Dependencies
sphinx-autoapi>=3.0.0
sphinx-autodoc-typehints>=1.24.0

# Optional: GPU Support for ML (uncomment if needed)
# torch>=2.1.0+cu118
# torchvision>=0.16.0+cu118
# tensorflow-gpu>=2.13.0